Move the theme root to the components folder
